backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery==5.5.3",
    "channels==4.3.1",
    "channels-redis==4.3.0",
    "cryptography>=45.0.7",
    "django>=4.0,<5.0",
    "django-apscheduler==0.7.0",
    "django-cors-headers==4.7.0",
    "django-environ>=0.12.0",
    "django-filter==25.1",
    "django-redis==6.0.0",
    "django-user-agents==0.4.0",
    "djangorestframework==3.16.1",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-writable-nested==0.7.2",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "pycryptodome==3.23.0",
    "pymysql>=1.1.2",
    "redis==6.4.0",
    "requests>=2.32.5",
]


[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
